Refactor runtime helpers into modules
