+++ src/runtime_opcode.h
@@ -1,0 +1,9 @@
+// Runtime opcode declarations and dispatch binding
+
+// Licensed under the MIT License - see LICENSE file for details.
+
+#pragma once
+
+#include "runtime_context.h"
+
+void runtime_opcode_bind_table(RuntimeContext *ctx);
